CORE-0014 Fake objects now in BaseTest, test maps deleted
